requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
python-socketio==5.11.0
pyserial==3.5
mysql-connector-python==8.2.0
requests==2.31.0
